Split pdf service to interface and implementation
